Updated Word engine and tests
Updated the word engine to check to make sure the first and last word are the same length and added tests.

Checked to make sure the words exists in the list. Need to add tests for this.

First attempt at finding words. It currently replaces each letter at a time in order. If the word is not found it throws and expection